IDVA6-1180 add welsh translations and fix formatting
