+++ src/lib/validation/add.user.validation.ts
@@ -1,5 +1,6 @@
 import { ViewData } from "../../types/utilTypes";
 import { validateEmailString } from "../../lib/validation/email.validation";
+import { addErrorToViewData } from "../../lib/utils/viewUtils";
 import { validateRole } from "./user.role.validation";
 import * as constants from "../constants";
 
@@ -7,19 +8,6 @@ const FormInputNames = {
     EMAIL: "email",
     USER_ROLE: "userRole"
 } as const;
-
-export const addErrorToViewData = (
-    errProp: string,
-    errorMsg: string,
-    viewData: ViewData
-): void => {
-    viewData.errors = {
-        ...viewData.errors,
-        [errProp]: {
-            text: errorMsg
-        }
-    };
-};
 
 export const validateAndSetErrors = (
     email: string,
@@ -36,7 +24,7 @@ export const validateAndSetErrors = (
     }
 };
 
-export const validateAddUserEmail = (email:string):string|undefined => {
+export const validateAddUserEmail = (email: string): string | undefined => {
     if (!email) {
         return constants.ERRORS_EMAIL_REQUIRED;
     } else if (!validateEmailString(email)) {
