+++ src/lib/validation/user.role.validation.ts
@@ -1,8 +1,12 @@
 import * as constants from "../constants";
 import { UserRole } from "private-api-sdk-node/dist/services/acsp-manage-users/types";
 
-export const validateRole = (role:string|undefined):string|undefined => {
-    if (role === UserRole.OWNER || role === UserRole.ADMIN || role === UserRole.STANDARD) {
+export const validateRole = (role: string | undefined): string | undefined => {
+    if (
+        role === UserRole.OWNER ||
+        role === UserRole.ADMIN ||
+        role === UserRole.STANDARD
+    ) {
         return undefined;
     }
     return constants.ERRORS_SELECT_USER_ROLE;
