+++ src/locales/cy/translation/add-user.json
@@ -1,0 +1,24 @@
+{
+    "page_header":  "[CY] Add a user",
+    "back_link_to_manage_users":  "[CY] Return to manage users",
+    "p1":  "[CY] In the future, users who have been added to this account will be able to:",
+    "bullet_1":  "[CY] file as an authorised agent",
+    "bullet_2":  "[CY] verify people's identity for Companies House",
+    "email_input_label":  "[CY] Email address",
+    "email_hint_text":  "[CY] This must be the email address they use for their Companies House account. If they do not have a Companies House account, they must create one before you can add them.",
+    "radio_btns_header":  "[CY] What is their role?",
+    "option_1":  "[CY] Account owner",
+    "option_1_hint_p":  "[CY] They will be able to make changes to the authorised agent's details. They can add and remove:",
+    "option_1_hint_bullet_1":  "[CY] account owners",
+    "option_1_hint_bullet_2":  "[CY] administrators",
+    "option_1_hint_bullet_3":  "[CY] standard users",
+    "option_2":  "[CY] Administrator",
+    "option_2_hint_p":  "[CY] They will be able to add and remove:",
+    "option_2_hint_bullet_1":  "[CY] administrators",
+    "option_2_hint_bullet_2":  "[CY] standard users",
+    "option_3":  "[CY] Standard user",
+    "option_3_hint":  "[CY] They will not be able to add or remove other users.",
+    "errors_email_required":  "[CY] Enter an email address",
+    "errors_email_invalid":  "[CY] Enter an email address in the correct format, like name@example.com",
+    "errors_select_user_role":  "[CY] Select the user’s role"
+}
