+++ src/locales/cy/translation/common.json
@@ -3,5 +3,8 @@
     "skip_to_main_content": "[CY] Skip to main content",
     "there_is_a_problem": "[CY] There is a problem",
     "customer_insights": "[CY] Customer insights",
-    "back_link_to_the_previous_page": "[CY] Backlink to the previous page"
+    "back_link_to_the_previous_page": "[CY] Backlink to the previous page",
+    "continue": "[CY] Continue",
+    "continue_button_to_proceed_to_the_next_page": "[CY] Continue button to proceed to the next page",
+    "title_error": "[CY] Error: "
 }
